modified:   CSCI 201/Final Project/Budget.cpp 	modified:   CSCI 201/Final Project/Budget.h 	modified:   CSCI 201/Final Project/People.h 	new file:   CSCI 201/Final Project/ProjectSummary.docx 	modified:   CSCI 201/Final Project/Restaurant.cpp 	modified:   CSCI 201/Final Project/Restaurant.h 	modified:   CSCI 201/Final Project/RestaurantSimulator 	new file:   CSCI 201/Final Project/SampleSubmission.docx 	modified:   CSCI 201/Final Project/Task.cpp 	modified:   CSCI 201/Final Project/Task.h 	modified:   CSCI 201/Final Project/main.cpp

+++ CSCI 201/Final Project/People.h
@@ -30,7 +30,7 @@ struct menuItem{
     string itemName;
     double itemPrice;
     double ingredientCost;
-    double timeToMake; // Double between 0.5 and 1.5
+    int timeToMake; // Minutes between 10 and 30
 };
